+++ py/runTunnel.py
@@ -12,10 +12,7 @@ class RunThreeVScript(ThreeVScript.ThreeVScript):
 	#=====================
 	def setupParserOptions(self):
 		self.parser.set_usage("Usage: %prog --rundir=<output dir> --jobid=<job id> --pdbid=<pdb id>")
-		self.parser.add_option("--bigprobe", dest="bigprobe", type="float", default=6.0,
-			help="Probe radius", metavar="#")
-		self.parser.add_option("--smallprobe", dest="smallprobe", type="float", default=2.0,
-			help="Probe radius", metavar="#")
+		ThreeVScript.add_dual_probe_options(self.parser)
 
 	#=====================
 	def checkConflicts(self):
@@ -35,14 +32,7 @@ class RunThreeVScript(ThreeVScript.ThreeVScript):
 		#giffile = self.threev.animateGif(pngfiles)
 
 		### write to webpage
-		f = open("results-"+self.params['jobid']+".html", "w")
-
-		self.threev.webImageSection(pngfiles, self.website, f)
-		self.threev.webJmolSection(objfile, self.website, f)
-		self.threev.webMrcStats(mrcfile, self.params['gridsize'], f)
-		self.threev.webMrcSection([mrcfile], self.website, f, pdb=True, pymol=self.params['pymol'])
-
-		f.close()
+		self.write_single_mrc_results_page(mrcfile, pngfiles, objfile)